vim(mappings/navigation): use `:e#` instead of `:b#`
Open a new buffer if the file that we are switching to is currently not
open in a buffer (e.g. if it was previously closed via `:bd`).

Previously, it was reusing the same buffer, replacing the file that
buffer was originally meant for.

One caveat: `:e#` cannot switch back to non-existent files. For
instance, `[No Name]` buffers.